+++ modules/module-mongodb/src/storage/implementation/MongoSyncBucketStorage.ts
@@ -237,7 +237,7 @@ export class MongoSyncBucketStorage
         },
         {
           $group: {
-            _id: '$key',
+            _id: { key: '$key', lookup: '$lookup' },
             bucket_parameters: {
               $first: '$bucket_parameters'
             }
@@ -494,11 +494,13 @@ export class MongoSyncBucketStorage
     while (true) {
       try {
         await this.clearIteration();
+
+        logger.info(`${this.slot_name} Done clearing data`);
         return;
       } catch (e: unknown) {
         if (e instanceof mongo.MongoServerError && e.codeName == 'MaxTimeMSExpired') {
           logger.info(
-            `Clearing took longer than ${db.mongo.MONGO_CLEAR_OPERATION_TIMEOUT_MS}ms, waiting and triggering another iteration.`
+            `${this.slot_name} Cleared batch of data in ${db.mongo.MONGO_CLEAR_OPERATION_TIMEOUT_MS}ms, continuing...`
           );
           await timers.setTimeout(db.mongo.MONGO_CLEAR_OPERATION_TIMEOUT_MS / 5);
           continue;