+++ modules/module-mysql/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @powersync/service-module-mysql
 
+## 0.1.6
+
+### Patch Changes
+
+- Updated dependencies [320e646]
+- Updated dependencies [e3a9343]
+  - @powersync/service-core@0.12.2
+
+## 0.1.5
+
+### Patch Changes
+
+- Updated dependencies [889ac46]
+  - @powersync/service-core@0.12.1
+
 ## 0.1.4
 
 ### Patch Changes